requirements.txt
allure-python-commons
selenium
pytest==8.4.2
selene == 2.0.0rc9
allure-pytest==2.15.0
python-dotenv==0.20.0